# packages/backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistem POS API",
    "version": "1.0.0",
    "description": "Dokumentasi REST API Sistem POS real-time multi-tenant.",
    "contact": {
      "name": "Sistem POS Team",
      "email": "support@sistempos.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Terjadi kesalahan"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login pengguna",
        "description": "Melakukan autentikasi pengguna dan mengembalikan access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password",
                  "tenantId"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username pengguna",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password pengguna",
                    "example": "admin123"
                  },
                  "tenantId": {
                    "type": "string",
                    "description": "ID tenant",
                    "example": "7f69ce68-9068-11f0-8eff-00155d24a169"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Login successful"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "description": "Data pengguna"
                        },
                        "accessToken": {
                          "type": "string",
                          "description": "JWT access token"
                        },
                        "refreshToken": {
                          "type": "string",
                          "description": "JWT refresh token"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Kredensial tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Akses tenant tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh access token",
        "description": "Memperbarui access token menggunakan refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "JWT refresh token",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Token refreshed successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "description": "JWT access token baru"
                        },
                        "refreshToken": {
                          "type": "string",
                          "description": "JWT refresh token baru"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh token tidak valid atau expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}